forked player scene into player_prop and player_hunter
prop is largely the same as before with a slight improvment to the 'attack' event handling.
hunter is a clone of prop with the body-changing mechanics removed.
the hunter attack has been replaced with placeholder behavior.
hunter retains the Q camera lock behavior since that code is on the camera orbit.
the camera lock functionality should be migrated to player_prop so that it is not shared across player types.