fixed bug in remote player_prop collision shape updates
the rpc for updating player_prop puppets was attempting to send collision shapes over the network. This does not work.
now, a string is passed which references a prop rigidBody from which to copy collision shapes from.
this, coincidentally, also fixes a bug where player-on-player collisions between one or more player_props would cause both players to become stuck.